Compat with ESP-IDF master

+++ src/i2s.rs
@@ -103,9 +103,10 @@ pub mod config {
 
     impl ClockSource {
         #[cfg(not(esp_idf_version_major = "4"))]
+        #[allow(clippy::unnecessary_cast)]
         pub(super) fn as_sdk(&self) -> i2s_clock_src_t {
             match self {
-                Self::Pll160M => esp_idf_sys::soc_module_clk_t_SOC_MOD_CLK_PLL_F160M,
+                Self::Pll160M => esp_idf_sys::soc_module_clk_t_SOC_MOD_CLK_PLL_F160M as u32,
                 #[cfg(any(esp32, esp32s2))]
                 Self::Apll => esp_idf_sys::soc_module_clk_t_SOC_MOD_CLK_APLL,
             }
